[charts] Split axis utils from main file
